Create ja.java
I tried multiple times trying to upload the .class executable to this site, but got informed of the "file type error" every time. I thought I had come to the right place, because that was the concept "release" all about in here, right? What you try to release should be binary, executble stuff out of box, or am I misled again?

So I upload the source code here instead, you can compile it in your machine and put the .class file in the proper path. As to how to use it in the command line environment, read the java source code and you get it all, it's just a bunch of lines. For me it's remarkably useful and equally remarkably simple. I had been quite upset with the shipped calculator in my Win8.1, quite a pain in where you sit on for me to use. And I've grown tired of bothering with the mouse clicking all along on the little piece of window.

Anyone interested is welcome to fork it and improve it, as long as keep it simple and handy to use on a daily basis. Happy forking!

Leo